Caching getMembers
